+++ test/fixtures/outside-click-hook/src/useOutsideClickAlerter.ts
@@ -1,0 +1,15 @@
+import { type RefObject, useEffect } from "react"
+
+export default function useOutsideClickAlerter(ref: RefObject<HTMLElement>, callback: () => void) {
+  useEffect(() => {
+    function handleClickOutside(event: MouseEvent) {
+      if (ref.current && !ref.current.contains(event.target as HTMLElement)) {
+        callback()
+      }
+    }
+    document.addEventListener("mousedown", handleClickOutside)
+    return () => {
+      document.removeEventListener("mousedown", handleClickOutside)
+    }
+  }, [ref, callback])
+}